requirements.txt
matplotlib
numpy
pandas
pillow
pytorch
scikit-learn
scipy
tqdm
transformers
torchvision
tqdm
umap-learn